Implementation of new bad URI handling logic.
This changeset is intended to resolve
aarongustafson/jekyll-webmention_io#129.  Here we implement a much more
flexible policy mechanism for handling URIs for which the plugin
previous experienced errors sending a webmention.

Errors are classified into three different categories (unsupported,
error, failure), and separate policies can be defined for each, allowing
the user a lot of control over how the plugin treats these conditions.

Additionally, while the plugin currently has a feature to control
retries for whole hosts, what it lacks is a mechanism to have the plugin
give up on individual webmentions.

This commit introduces a new max_attempts setting which, if set, will
cause the plugin to give up and mark a webmention as sent after a
configured number of tries.