Introduce max_attempts setting.
Where the plugin currently has a feature to control retries for whole
hosts, what it lacks is a mechanism to have the plugin give up on
individual webmentions.  This is really another aspect of issue #129.

This commit introduces a new max_attempts setting which, if set, will
cause the plugin to give up and mark a webmention as sent after a
configured number of tries.