Add configurable syndication response handling.
This commit actually does a couple of things.

First, the configuration has been restructured to allow for a bit more
flexibility in how syndication endpoints are configured.

We then add support for defining a mapping between the responses from
syndication webmention endpoints and the page front matter.

This eliminates the hardcoded handling for Brid.gy specifically.