Support for shorturls.
This commit alters the plugin so that, if a page has a shorturl defined
in its front matter, that URL will be used as the webmention source
instead of the authoritative URL.

Among other things, this means that when you use the plugin with
brid.gy, it sends over the short URL to places like Twitter.

This behaviour is configurable by syndication target.