Fix file creation when no JS file is requested
The refactoring of the file creation code to fix the bug introduced with
Jekyll 4 lead to a different issue, where the file inside the temporary
directory is now potentially nested in a subdirectory.

Rather than making the code more complex, I just switched to mkdir_p, so
the directory is created recursively.