Do more work even if we're paused.
A key part of the syndication functionality is to pull back the
permalink from the syndicated post and stick that in the frontmatter so
the link can be embedded in the post when the site is built.

However, if we pause gathering/queuing mentions, we still need to do
that for a site build to come out looking right.

So, when queuing outgoing mentions (which is also, right now, the same
codepath where that permalink storage happens), the pause setting now
just controls whether we write out the new cache or not.  We still do
most of the work (which is mainly parsing pages for URLs and so forth).