Rework of webmention templates
This commit is intended to address issues #175, #176, and #177.  It is a
fairly involved reworking of these templates that standardizes the
layout and, I hope, does a better job of rendering h-card information
for posts and so forth.

These changes introduce some structural changes to the HTML, including
consistently embedding webmention content in article tags, and
repositioning some of the relevant classes.

Any users of the default templates will want to inspect these changes
and make corresponding changes to their CSS.

Now, I have to admit, the sheer amount of code duplication across these
templates fills me with sadness, but because of the need to support both
client- and server-side template rendering, these things kinda need to
stand on their own.  I did explore the use of custom filters to render
blocks like the h-card using their own templates, but figuring out how
to make a custom filter on the JS side, not to mention testing it, was
beyond my ken.