Implement html_proofer_ignore setting
This changeset is intended to address issue #172.  It introduces a new
html_proofer_ignore setting which can take the values "templates" or
"all".  When set to "templates", the behaviour is equivalent to setting
the html_proofer setting to true.  When set to "all", in addition to
that original behaviour, the included webmention templates will add the
data-proofer-ignore attribute to the rendered webmention output, which
will cause any content sent via webmention to be ignored by the proofer.