Reject malformed URIs during gather phase
This commit addresses issue #178 by adding an additional check during
the URI gathering phase such that, after the initial regex match finds a
URI, we attempt to parse that URI.  If that fails, we ignore the URI as
a spurious match and move on.

This is, of course, a hack.  In practice we should clean up that regex
so that bad matches don't happen.  But, since I don't know enough about
the history of that regex, and because that could turn into a game of
whack-a-mole, I went this route.  And, ultimately, this is just more
bullet-proof.

Note, these events are silently ignored at the regular logging level.
However, at debug level, a log message is emitted indicating a URI was
skipped.