Always set_data when calling render_into_template
This commit fixes issue #179. The method `set_data` sets the class
instance variable @data with essential state information needed during
the render call.

However, as this tag is instantiated during the creation of a template,
its instance persists if the template is rendered multiple times. So if
a page is rendered with this template, and that page contains
webmentions, and then a subsequent page is rendered that has no
webmentions, if that state isn't cleared out, then it's erroneously used
in that subsequent call.

Well, prior to this changeset, the `set_data` method was only called if
webmentions were available for a given page. That meant pages without
any webmentions failed to clear the previously cached state, and the
webmentions would end up re-rendered in the current page.

This changeset modifies this method (and cleans up the ordering of the
code a bit) so that the state is always populated, regardless of the
presence of webmentions on the page. This ensures the @data variable
always contains the state for the current page being rendered, even if
that state includes an empty list of webmentions.