Add a note about the changelog being retired
Separate changelogs have always bothered me. As a document targeting a
technical audience, they become a burden to maintain and ensure they're
up to date. Meanwhile, the entire purpose of a commit history is to
document the changes to a project over time.

The main reason changelogs are nice is they allow one to summarize those
changes, but clean and very well documented commits are, at least in my
opinion, a far better solution for small projects where a deeply
opinionated maintainer can ensure commit descriptions meet the necessary
standard, as it avoids duplication of information while raising the
standard for the source of truth: the actual commit history.