Lift out caching logic into a contained module
Ultimately the goal is to simplify the bootstrap method of
Jekyll::WebmentionIO to take a set of service dependencies, and then
reorganize or eliminate any core dependencies into said services.

As an easy first step, here, we take all the caching code and lift it
out into a service module that components can initialize directly.