Move policy logic into service class
Next phase of refactoring, now we move a bunch of policy-related code
into a self-contained module. This is a *slightly* weird refactoring in
that I've put the post throttling logic in with the URI handling logic,
but since they're both policy related it seemed... okay? Not ideal, but
it seemed overkill to throw that one bit of logic into its own class.