Moved gather_documents into config class
I was a little torn about where to shove this, but given the
gather_documents implementation is so tied to the configuration, the
Config class didn't seem... totally wrong?

This last change removes the last bit of core logic out of
webmention_io.rb with the one exception of the log() method, which I'm
going to leave in place for now, as well as a global webmention type
list that could really live anywhere, so why not there?

Now I guess I should actually write some unit tests...