Rename "next lookup" to "lookup threshold"
After better understanding how the throttling code works, I realized
what's really happening is the code is calculating a threshold date such
that if the previous lookup is *before* that threshold, then we're
allowed to perform another lookup. So I wanted to clarify the method
naming a bit to make that more obvious.