Refactor existing test harness code
This changeset is in prep for adding the first round of new unit tests.
Here we refactor the code in `spec_helper.rb` to throw it in a namespace
and create some necessary mocks, then update the current unit test code
to work.

And in the process I also killed a broken test that hasn't worked since
at least v4.0.0. I'll have to go back and revisit that later...