Fixed another refactoring bug
This was so broken for so long... I had some code commented out in my
working tree and didn't notice the breakage until I reverted those
changes. Woops!

I'd rebase this fix and squash it into the earlier refactoring but it's
too much work so we'll just live with the wart in the commit history...