Major refactoring and new unit test suite (#183)
This is a very large changeset that substantially refactors quite a bit
of the code in order to move a lot of shared logic into service classes,
introduces a simple DI pattern for making those services available to
classes that need them, and then adds a whole raft of unit tests to
validate the whole thing while taking advantage of this new, more
testable design.

This does come with a few behavioural changes here and there related to
logging output, but those changes should be entirely cosmetic.
Semantically, the code should behave as it did prior to these changes.

If I'm being honest, this service/DI approach isn't my absolute
favourite, but it was the least disruptive way to make these changes
without expanding the blast radius even further.